Added e-mail notification to fundraiser owner + unittest
