+++ apps/donations/templates/new_oneoff_donation_fundraiser.mail.html
@@ -1,0 +1,15 @@
+{% extends "base.mail.html" %}
+{% load i18n %}
+
+{% block content %}
+{% blocktrans with amount=amount donor_name=donor_name site=site link=link %}
+Hello,<br/>
+<br/>
+You just received a EUR {{ amount|floatformat:2 }} donation from {{ donor_name }}!<br/>
+<br/>
+Quickly thank your supporter by posting a personal message on your <a href="{{ site }}{{ link }}">fundraise action wall</a>.<br/>
+<br/>
+Cheers,<br/>
+1%Club
+{% endblocktrans %}
+{% endblock %}
