+++ apps/donations/templates/new_oneoff_donation_fundraiser.mail.txt
@@ -1,0 +1,8 @@
+Hello,
+
+You just received a EUR {{ amount|floatformat:2 }} donation from {{ donor_name }}!
+
+Quickly thank your supporter by posting a personal message on your fundraise action wall: {{ site }}{{ link }}.
+
+Cheers,
+1%Club
